scraper/requirements.txt
playwright==1.50.0
httpx==0.28.1
beautifulsoup4==4.13.3
supabase==2.13.0
python-dotenv==1.0.1
lxml==5.3.1
